Changed name and structure of absolutely everything, bro.
Makes it easier to find stuff.
