+++ src/civilization/census/census_components.rs
@@ -1,9 +1,9 @@
 use bevy::prelude::{Component, Reflect};
 
-#[derive(Component, Debug, Reflect)]
+#[derive(Component, Debug, Reflect, Default)]
 pub struct Census {
     pub population: usize,
 }
 
-#[derive(Component, Debug, Reflect)]
+#[derive(Component, Debug, Reflect, Default)]
 pub struct HasPopulation;
