+++ src/civilization/city_construction/mod.rs
@@ -1,4 +1,4 @@
-pub mod plugin;
-pub mod components;
-mod systems;
-pub mod events;
+pub mod city_construction_plugin;
+pub mod city_construction_components;
+mod city_construction_systems;
+pub mod city_construction_events;
