+++ src/civilization/city_support/city_support_components.rs
@@ -1,0 +1,8 @@
+use bevy::prelude::{Component, Reflect};
+
+#[derive(Component, Debug, Reflect)]
+pub struct HasTooManyCities {
+    pub surplus_count: usize,
+    pub needed_tokens: usize
+}
+
