+++ src/civilization/city_support/mod.rs
@@ -1,1 +1,4 @@
-pub mod plugin;
+pub mod city_support_plugin;
+pub mod city_support_components;
+pub mod city_support_events;
+pub mod city_support_systems;
