+++ src/civilization/conflict/conflict_components.rs
