+++ src/civilization/conflict/mod.rs
@@ -1,3 +1,3 @@
-pub mod plugin;
-pub mod systems;
-pub mod components;
+pub mod conflict_plugin;
+pub mod conflict_systems;
+pub mod conflict_components;
