+++ src/civilization/console/mod.rs
@@ -1,1 +1,1 @@
-pub mod commands;
+pub mod console_commands;
