--- src/civilization/game_phases/mod.rs
@@ -1,3 +1,0 @@
-pub mod systems;
-pub mod plugin;
-pub mod game_activity;
