--- src/civilization/general/enums.rs
@@ -1,7 +1,0 @@
-use bevy::prelude::Reflect;
-
-#[derive(Debug, Reflect, Copy, Clone, Eq, PartialEq)]
-pub enum GameFaction {
-    Egypt,
-    Crete,
-}
