+++ src/civilization/general/general_enums.rs
@@ -1,0 +1,9 @@
+use bevy::prelude::Reflect;
+
+#[derive(Debug, Reflect, Copy, Clone, Eq, PartialEq, Default)]
+pub enum GameFaction {
+    #[default]
+    Egypt,
+    Crete,
+}
+
