+++ src/civilization/mod.rs
@@ -1,6 +1,5 @@
 pub mod census;
 pub mod general;
-pub mod game_phases;
 pub mod population_expansion;
 pub mod movement;
 pub mod console;
