+++ src/civilization/population_expansion/mod.rs
@@ -1,4 +1,4 @@
-pub mod plugin;
-pub mod systems;
-pub mod components;
+pub mod movement_plugin;
+pub mod movement_systems;
+pub mod movement_components;
 pub mod events;
