+++ src/civilization/population_expansion/movement_components.rs
