+++ src/lib.rs
@@ -13,7 +13,7 @@ use crate::audio::InternalAudioPlugin;
 use crate::loading::LoadingPlugin;
 use crate::menu::MenuPlugin;
 
-use crate::civilization::general::plugin::CivilizationPlugin;
+use crate::civilization::general::general_plugin::CivilizationPlugin;
 use bevy::app::App;
 // #[cfg(debug_assertions)]
 // use bevy::diagnostic::{FrameTimeDiagnosticsPlugin, LogDiagnosticsPlugin};
@@ -28,6 +28,26 @@ pub enum GameState {
     Playing,
     // Here the menu is drawn and waiting for player interaction
     Menu,
+}
+
+#[derive(SubStates, Clone, PartialEq, Eq, Hash, Debug, Default)]
+#[source(GameState = GameState::Playing)]
+pub enum GameActivity {
+    // CollectTaxes,
+    #[default]
+    PopulationExpansion,
+    Census,
+    // ShipConstruction,
+    Movement,
+    Conflict,
+    CityConstruction,
+    RemoveSurplusPopulation,
+    CheckCitySupport,
+    // AcquireTradeCards,
+    // Trade,
+    // ResolveCalamities,
+    // AcquireCivilizationCards,
+    // MoveSuccessionMarkers,
 }
 
 pub struct GamePlugin;
